Add message on who won.
